+++ react/components/molecules/index.js
@@ -1,5 +1,6 @@
 export * from "./button-toggle";
 export * from "./chat-message";
+export * from "./checkbox-group";
 export * from "./date-input";
 export * from "./header";
 export * from "./input-animated";