+++ CHANGELOG.md
@@ -27,6 +27,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Textarea support for background color, horizontal padding and font size as props
 * Input keyboard type support
 * Input Form component
+* Possibility to override icon props in `tabs` and `button-tab`
+* Checkbox group component
+* Support for `Avatar`'s image change
+* Support for `View more` button in key-values
 
 ### Changed